// src/modules/auth/auth.enum.ts
export enum LogoutType {
  THIS_DEVICE = 'Thiết bị',
  ALL_DEVICES = 'Tất cả thiết bị',
}

export enum RefreshTokenStatus {
  USED = 'Used',
  UNUSED = 'Unused',
}

export enum ResetPasswordType {
  OTP = 'Otp',
  EMAIL = 'Email',
}

export enum ForgotPasswordType {
  PHONE = 'Phone',
  EMAIL = 'Email',
}